+++ cmd/binder/app/app.go
@@ -107,7 +107,7 @@ func New(options *Options, config *rest.Config) (*App, error) {
 	rrs := resourcereservation.NewService(options.FakeGPUNodes, clientWithWatch, options.ResourceReservationPodImage,
 		time.Duration(options.ResourceReservationAllocationTimeout)*time.Second,
 		options.ResourceReservationNamespace, options.ResourceReservationServiceAccount,
-		options.ResourceReservationAppLabel, options.ScalingPodNamespace)
+		options.ResourceReservationAppLabel, options.ScalingPodNamespace, options.RuntimeClassName)
 
 	reconcilerParams := &controllers.ReconcilerParams{
 		MaxConcurrentReconciles:     options.MaxConcurrentReconciles,