+++ cmd/binder/app/options.go
@@ -29,6 +29,7 @@ type Options struct {
 	FakeGPUNodes                         bool
 	GpuCdiEnabled                        bool
 	VolumeBindingTimeoutSeconds          int
+	RuntimeClassName                     string
 }
 
 func InitOptions(fs *pflag.FlagSet) *Options {
@@ -93,6 +94,9 @@ func InitOptions(fs *pflag.FlagSet) *Options {
 	fs.IntVar(&options.VolumeBindingTimeoutSeconds,
 		"volume-binding-timeout-seconds", 120,
 		"Volume binding timeout in seconds")
+	fs.StringVar(&options.RuntimeClassName,
+		"runtime-class-name", "",
+		"Runtime class for reservation pods")
 
 	utilfeature.DefaultMutableFeatureGate.AddFlag(fs)