+++ pkg/apis/kai/v1/binder/binder.go
@@ -99,6 +99,9 @@ type ResourceReservation struct {
 	// AppLabel is the value that will be set for all resource reservation pods to the label `app`
 	// +kubebuilder:validation:Optional
 	AppLabel *string `json:"appLabel,omitempty"`
+
+	// RuntimeClassName specifies the runtime class used by the reservation pods. Needs to allow access to the GPU
+	RuntimeClassName *string `json:"runtimeClassName,omitempty"`
 }
 
 func (r *ResourceReservation) SetDefaultsWhereNeeded() {
@@ -109,4 +112,5 @@ func (r *ResourceReservation) SetDefaultsWhereNeeded() {
 	r.Namespace = common.SetDefault(r.Namespace, ptr.To(constants.DefaultResourceReservationName))
 	r.ServiceAccountName = common.SetDefault(r.ServiceAccountName, ptr.To(constants.DefaultResourceReservationName))
 	r.AppLabel = common.SetDefault(r.AppLabel, ptr.To(constants.DefaultResourceReservationName))
+	r.RuntimeClassName = common.SetDefault(r.RuntimeClassName, ptr.To(constants.DefaultRuntimeClassName))
 }