+++ pkg/common/constants/constants.go
@@ -20,6 +20,7 @@ const (
 	DefaultScaleAdjustName                = "kai-scale-adjust"
 	DefaultKAIConfigSingeltonInstanceName = "kai-config"
 	DefaultNodePoolLabelKey               = "kai.scheduler/node-pool"
+	DefaultRuntimeClassName               = "nvidia"
 
 	DefaultCPUWorkerNodeLabelKey = "node-role.kubernetes.io/cpu-worker"
 	DefaultGPUWorkerNodeLabelKey = "node-role.kubernetes.io/gpu-worker"